requirements.txt
numpy==1.19.3
matplotlib==3.4.3
scikit-learn==1.0
sympy==1.9
scikit-image==0.18.3
scipy==1.7.1
tqdm==4.62.3
Pillow~=8.4.0